requirements.txt
Django>=4.2.1
Pillow
django-allauth
django-crispy-forms
crispy_bootstrap5
python-dotenv
requests
jwt
libscrc 
qrcode
psycopg2-binary